Company registration updates

+++ shuup_tests/functional/test_notify_on_company_created.py
@@ -10,6 +10,7 @@ import pytest
 from django.conf import settings
 from django.core.urlresolvers import reverse
 
+from shuup import configuration
 from shuup.core.models import get_company_contact, get_person_contact
 from shuup.front.apps.customer_information.notify_events import CompanyAccountCreated
 from shuup.notify.actions.notification import AddNotification
@@ -24,14 +25,15 @@ from shuup_tests.utils.fixtures import (
 )
 
 
-
 @pytest.mark.django_db
-def test_notify_on_company_created(regular_user):
+@pytest.mark.parametrize("allow_company_registration", (False, True))
+def test_notify_on_company_created(regular_user, allow_company_registration):
     if "shuup.front.apps.customer_information" not in settings.INSTALLED_APPS:
         pytest.skip("shuup.front.apps.customer_information required in installed apps")
     if "shuup.notify" not in settings.INSTALLED_APPS:
         pytest.skip("shuup.notify required in installed apps")
 
+    configuration.set(None, "allow_company_registration", allow_company_registration)
     step = Step(
         cond_op=StepConditionOperator.NONE,
         actions=[
@@ -55,26 +57,32 @@ def test_notify_on_company_created(regular_user):
     client = SmartClient()
     client.login(username=REGULAR_USER_USERNAME, password=REGULAR_USER_PASSWORD)
     company_edit_url = reverse("shuup:company_edit")
-    client.soup(company_edit_url)
 
-    data = _default_company_data()
-    data.update(_default_address_data("billing"))
-    data.update(_default_address_data("shipping"))
+    if allow_company_registration:
+        client.soup(company_edit_url)
 
-    response, soup = client.response_and_soup(company_edit_url, data, "post")
+        data = _default_company_data()
+        data.update(_default_address_data("billing"))
+        data.update(_default_address_data("shipping"))
 
-    assert response.status_code == 302
-    assert get_company_contact(regular_user)
-    assert Notification.objects.filter(identifier="company_created").count() == 1
-    notification = Notification.objects.filter(identifier="company_created").first()
-    assert notification
-    assert data["contact-email"] in notification.message
+        response, soup = client.response_and_soup(company_edit_url, data, "post")
 
-    # New save should not add new notifications
-    response, soup = client.response_and_soup(company_edit_url, data, "post")
-    assert response.status_code == 302
-    assert Notification.objects.filter(identifier="company_created").count() == 1
-    script.delete()
+        assert response.status_code == 302
+        assert get_company_contact(regular_user)
+        assert Notification.objects.filter(identifier="company_created").count() == 1
+        notification = Notification.objects.filter(identifier="company_created").first()
+        assert notification
+        assert data["contact-email"] in notification.message
+
+        # New save should not add new notifications
+        response, soup = client.response_and_soup(company_edit_url, data, "post")
+        assert response.status_code == 302
+        assert Notification.objects.filter(identifier="company_created").count() == 1
+        script.delete()
+    else:
+        response = client.get(company_edit_url)
+        assert response.status_code == 404
+        assert Notification.objects.filter(identifier="company_created").count() == 0
 
 
 def _default_company_data():
